[core] feat: separar núcleo e interface CLI
